// admin/src/Components/AdminTable/Filters/Amount/interfaces.ts
import { Column } from '@tanstack/react-table';
import { ColumnFilters } from 'Components/AdminTable/interfaces';
import { IPageListTableRow } from 'Hooks/useAdminTable/interfaces';

import { AMOUNT_FILTER } from './constants';

export interface IAmountFilterProps<T extends IPageListTableRow> {
	column: Column<T>;
	value: string;
	onClick?: (
		event: React.MouseEvent<HTMLSelectElement | HTMLInputElement>
	) => void;
	setColumnFilters: React.Dispatch<React.SetStateAction<ColumnFilters>>;
	setDataFilter: (column: Column<T>, value: string) => void;
}

export interface IAvailableCountFilter {
	comparingNumberType?: AMOUNT_FILTER;
	value?: number | '';
}